freq error can be float

+++ src/main/java/ru/r2cloud/lora/r2lora/R2loraClient.java
@@ -177,7 +177,7 @@ public class R2loraClient {
 		JsonObject obj = val.asObject();
 		LoraFrame result = new LoraFrame();
 		result.setData(Util.hexStringToByteArray(obj.getString("data", null)));
-		result.setFrequencyError(obj.getLong("frequencyError", 0));
+		result.setFrequencyError((long) obj.getFloat("frequencyError", 0));
 		result.setRssi((short) obj.getInt("rssi", 0));
 		result.setSnr(obj.getFloat("snr", 0));
 		result.setTimestamp(obj.getLong("timestamp", 0));
